+++ packages/nimble-components/src/table-column/number-text/tests/table-column-number-text.spec.ts
@@ -1,4 +1,5 @@
 import { html, ref } from '@microsoft/fast-element';
+import { parameterizeSpec } from '@ni/jasmine-parameterized';
 import { tableTag, type Table } from '../../../table';
 import { TableColumnNumberText, tableColumnNumberTextTag } from '..';
 import { waitForUpdatesAsync } from '../../../testing/async-helpers';
@@ -7,7 +8,6 @@ import type { TableRecord } from '../../../table/types';
 import { TablePageObject } from '../../../table/testing/table.pageobject';
 import { NumberTextAlignment, NumberTextFormat } from '../types';
 import type { TableColumnNumberTextCellView } from '../cell-view';
-import { parameterizeNamedList } from '../../../utilities/tests/parameterized';
 import { TextCellViewBaseAlignment } from '../../text-base/cell-view/types';
 import { lang, themeProviderTag } from '../../../theme-provider';
 import { unitByteTag } from '../../../unit/byte';
@@ -87,7 +87,7 @@ describe('TableColumnNumberText', () => {
             data: [{ number1: 'hello world' as unknown as number }]
         }
     ] as const;
-    parameterizeNamedList(noValueData, (spec, name, value) => {
+    parameterizeSpec(noValueData, (spec, name, value) => {
         spec(`displays empty string when ${name}`, async () => {
             await table.setData(value.data);
             await connect();
@@ -656,7 +656,7 @@ describe('TableColumnNumberText', () => {
         }
     ] as const;
     describe('sets the correct initial alignment on the cell', () => {
-        parameterizeNamedList(alignmentTestCases, (spec, name, value) => {
+        parameterizeSpec(alignmentTestCases, (spec, name, value) => {
             spec(name, async () => {
                 await table.setData([{ number1: 10 }]);
                 elementReferences.column1.format = value.format;