+++ packages/nimble-components/src/table/models/selection-managers/selection-manager-base.ts
@@ -41,19 +41,21 @@ export abstract class SelectionManagerBase<TData extends TableRecord> {
         isSelecting?: boolean
     ): void {
         if (
-            rowState.isGrouped
+            rowState.isGroupRow
             && rowState.selectionState === TableRowSelectionState.selected
         ) {
             // Work around for https://github.com/TanStack/table/issues/4759
             // Manually deselect all leaf rows when a fully selected group is being deselected.
             this.deselectAllLeafRows(rowState.id);
         } else {
-            this.tanStackTable.getRow(rowState.id).toggleSelected(isSelecting);
+            this.tanStackTable.getRow(rowState.id).toggleSelected(isSelecting, {
+                selectChildren: rowState.isGroupRow
+            });
         }
     }
 
     protected selectSingleRow(rowState: TableRowState): boolean {
-        if (rowState.isGrouped) {
+        if (rowState.isGroupRow) {
             throw new Error('function not intended to select grouped rows');
         }
 
@@ -101,7 +103,7 @@ export abstract class SelectionManagerBase<TData extends TableRecord> {
 
         return row
             .getLeafRows()
-            .filter(leafRow => leafRow.getLeafRows().length === 0)
+            .filter(leafRow => !leafRow.getIsGrouped())
             .map(leafRow => leafRow.id);
     }