+++ packages/nimble-components/src/text-area/tests/text-area.spec.ts
@@ -1,8 +1,8 @@
 import { html } from '@microsoft/fast-element';
+import { parameterizeSpec } from '@ni/jasmine-parameterized';
 import { TextArea, textAreaTag } from '..';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import { fixture, Fixture } from '../../utilities/tests/fixture';
-import { parameterizeNamedList } from '../../utilities/tests/parameterized';
 
 async function setup(): Promise<Fixture<TextArea>> {
     return fixture<TextArea>(html`<nimble-text-area></nimble-text-area>`);
@@ -79,7 +79,7 @@ describe('Text Area', () => {
         { name: 'aria-roledescription' }
     ] as const;
     describe('should reflect value to the internal control', () => {
-        parameterizeNamedList(attributeNames, (spec, name, value) => {
+        parameterizeSpec(attributeNames, (spec, name, value) => {
             spec(`for attribute ${name}`, async () => {
                 await connect();