+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,82 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Thu, 18 Jan 2024 17:21:22 GMT and should not be manually modified.
+This log was last generated on Mon, 29 Jan 2024 11:31:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 19.3.1
+
+Mon, 29 Jan 2024 11:31:46 GMT
+
+### Patches
+
+- Add validator APIs to rich text editor and viewer components ([ni/nimble@ec2e0dd](https://github.com/ni/nimble/commit/ec2e0ddef1d910d44097ccb2a976ad2343dbac63))
+- Bump @ni/nimble-components to v21.0.5
+
+## 19.3.0
+
+Sat, 27 Jan 2024 17:56:18 GMT
+
+### Minor changes
+
+- Adding Angular support for table data hierarchy ([ni/nimble@39bd1f8](https://github.com/ni/nimble/commit/39bd1f821cd5394c75187abbcf3c3dcc9b3a1572))
+
+## 19.2.4
+
+Thu, 25 Jan 2024 02:58:12 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v21.0.4
+
+## 19.2.3
+
+Wed, 24 Jan 2024 00:18:26 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v21.0.3
+
+## 19.2.2
+
+Tue, 23 Jan 2024 22:00:33 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v21.0.2
+
+## 19.2.1
+
+Tue, 23 Jan 2024 16:01:09 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v21.0.1
+
+## 19.2.0
+
+Mon, 22 Jan 2024 15:41:43 GMT
+
+### Minor changes
+
+- Create base mapping class in Angular ([ni/nimble@e91fdcd](https://github.com/ni/nimble/commit/e91fdcd3ece03166d22a148a291a72ce7a184c5d))
+
+## 19.1.0
+
+Fri, 19 Jan 2024 23:00:07 GMT
+
+### Minor changes
+
+- Angular support for nimble-unit-byte and nimble-unit-volt ([ni/nimble@07dfa53](https://github.com/ni/nimble/commit/07dfa53f1830b506e4a5f27c38ec879062d61eb1))
+
+## 19.0.0
+
+Fri, 19 Jan 2024 22:22:02 GMT
+
+### Major changes
+
+- Introducing data hierarchy for table. Renamed 'groupCollapseAll' on table label provider to 'collapseAll'. Update test with additional validation entry. ([ni/nimble@398b7ab](https://github.com/ni/nimble/commit/398b7ab7e2be2597de26e15e0ee463df4a02f228))
+- Bump @ni/nimble-components to v21.0.0
 
 ## 18.2.17